Move utils to src folder
Remove shared folder and move utils to src, as I think it is
a bit redundant to have the shared folder just for utils. If
we truly wanted a shared folder, then the config should also
be in the shared folder, as it is a shared resource.

In a future commit, the config file will be broken down into
smaller and modular config files. A "config" folder will be
created for this inside the "src" folder.